Criação de arquivos de utilitários para reutilzação
